requirements.txt
python-dotenv>=0.20.0,<1
nextcord>=2.1.0,<3
pandas>=1.5.3,<2